change saturation to a float

+++ src/main/java/parallelmc/parallelutils/modules/parallelitems/ParallelFish.java
@@ -4,4 +4,4 @@ import org.bukkit.inventory.ItemStack;
 
 import javax.annotation.Nullable;
 
-public record ParallelFish(int id, String key, int hunger, int saturation, @Nullable String cooked_key, ItemStack item) { }
+public record ParallelFish(int id, String key, int hunger, float saturation, @Nullable String cooked_key, ItemStack item) { }
